bug修改

+++ aries-base/src/main/java/com/cf/aries/base/service/impl/UserServiceImpl.java
@@ -29,7 +29,7 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public int insertUserInfo(UserInfo userInfo) {
+    public long insertUserInfo(UserInfo userInfo) {
         return userInfoMapper.insertSelective(userInfo);
     }
